chore: rename fileOperations.ts to file-operations.ts (with corresponding test-data/ file)
